refactor(test): create an lsp-specific helpers.lua file
* move the code for creating fake rpc servers there
